Calculator
3. Calculator 
Given two numbers a and b; you need to perform basic mathematical operation on them. You will be provided an integer named as operator. 

If operator equals to 1 add a and b, then print the result.

If operator equals to 2 subtract b from a, then print the result.

If operator equals to 3 multiply a and b, then print the result.

If operator equals to any another number, print "Invalid Input"(without quotes).